api/requirements/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
cssselect==1.2.0
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.6.1
et-xmlfile==1.1.0
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==5.1.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
parsel==1.8.1
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
pymongo==4.6.2
pyOpenSSL==24.0.0
python-dateutil==2.8.2
pytz==2024.1
queuelib==1.6.2
requests==2.31.0
requests-file==2.0.0
Scrapy==2.11.1
scrapy-djangoitem==1.1.1
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tldextract==5.1.1
Twisted==23.10.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
w3lib==2.1.2
zope.interface==6.2
